Add files via upload
The Zeyde code included a section where it looped over the test images and ran the bicubic scale up, the Yang dictionary, and their own pretrained dictionary on them. The zip file contains the results. The scale up factor is 3 for all of these. For some reason this comparison is converting all the images to B&W.

At the very least, this is something we can start to pull some results out of.